Rename plugin settings module to be more consistent
