+++ npm-packages/meteor-vite/src/VitePluginSettings.ts
