+++ npm-packages/meteor-vite/src/client/ValidateStub.ts
@@ -1,5 +1,5 @@
 import PackageJson from '../../package.json';
-import type { StubValidationSettings } from '../MeteorViteConfig';
+import type { StubValidationSettings } from '../VitePluginSettings';
 
 
 /**
