+++ npm-packages/meteor-vite/src/meteor/IPC/methods/vite-server.ts
@@ -1,7 +1,7 @@
 import FS from 'fs/promises';
 import Path from 'path';
 import { createServer, resolveConfig, ViteDevServer } from 'vite';
-import { MeteorViteConfig, type ProjectJson } from '../../../MeteorViteConfig';
+import { MeteorViteConfig, type ProjectJson } from '../../../VitePluginSettings';
 import { meteorWorker } from '../../../plugin/Meteor';
 import Logger from '../../../utilities/Logger';
 import { RefreshNeeded } from '../../../ViteLoadRequest';
