+++ npm-packages/meteor-vite/src/plugin/MeteorStubs.ts
@@ -8,7 +8,7 @@ import { createErrorHandler } from '../error/ErrorHandler';
 import { MeteorViteError } from '../error/MeteorViteError';
 import MeteorPackage from '../meteor/package/components/MeteorPackage';
 import { stubTemplate } from '../meteor/package/StubTemplate';
-import { MeteorViteConfig, type PluginSettings } from '../MeteorViteConfig';
+import { MeteorViteConfig, type PluginSettings } from '../VitePluginSettings';
 import ViteLoadRequest from '../ViteLoadRequest';
 
 export const MeteorStubs = setupPlugin(async () => {
